+++ src/Products/EasyNewsletter/tests/test_view_newsletter_issue_statistics.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+
+from plone import api
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from Products.EasyNewsletter.testing import PRODUCTS_EASYNEWSLETTER_FUNCTIONAL_TESTING
+from Products.EasyNewsletter.testing import PRODUCTS_EASYNEWSLETTER_INTEGRATION_TESTING
+from zope.component import getMultiAdapter
+from zope.component.interfaces import ComponentLookupError
+
+import unittest
+
+
+class ViewsIntegrationTest(unittest.TestCase):
+
+    layer = PRODUCTS_EASYNEWSLETTER_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.newsletter = api.content.create(
+            container=self.portal, type='Newsletter', id='newsletter',
+        )
+        self.issue = api.content.create(
+            container=self.newsletter, type='Newsletter Issue', id='issue',
+        )
+
+    def test_issue_statistics_is_registered(self):
+        view = getMultiAdapter(
+            (self.issue, self.portal.REQUEST),
+            name='statistics'
+        )
+        self.assertTrue(view.__name__ == 'statistics')
+
+    def test_issue_statistics_not_matching_interface(self):
+        with self.assertRaises(ComponentLookupError):
+            getMultiAdapter(
+                (self.portal, self.portal.REQUEST),
+                name='statistics'
+            )
+
+
+class ViewsFunctionalTest(unittest.TestCase):
+
+    layer = PRODUCTS_EASYNEWSLETTER_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])